(refactor): exibindo com sucesso produtos cadastrados

+++ src/main/java/br/com/edmilton/java/model/Produto.java
@@ -8,11 +8,11 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
-@Table(name="produtos")
+@Table(name = "produtos")
 public class Produto {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column
 	private String nome;
@@ -23,7 +23,7 @@ public class Produto {
 	@Column
 	private String descricao;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
@@ -39,7 +39,7 @@ public class Produto {
 		return descricao;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -54,5 +54,4 @@ public class Produto {
 	public void setDescricao(String descricao) {
 		this.descricao = descricao;
 	}
-
 }
